added convenience functions for loading files from local data
