+++ tests/helpers/__init__.py
@@ -1,10 +1,16 @@
+from .list_files_from_local_data import list_files_from_local_data
 from .load_files_from_local_data import load_files_from_local_data
+from .load_repo_files_from_local_data import load_repo_files_from_local_data
+from .load_user_files_from_local_data import load_user_files_from_local_data
 from .get_urls import get_urls
 from .skip_unreachable import skip_unreachable
 
 
 __all__ = [
     "get_urls",
+    "list_files_from_local_data",
     "load_files_from_local_data",
+    "load_repo_files_from_local_data",
+    "load_user_files_from_local_data",
     "skip_unreachable"
 ]
