+++ tests/helpers/load_files_from_local_data.py
@@ -1,4 +1,4 @@
-import os
+from .list_files_from_local_data import list_files_from_local_data
 
 
 def load_files_from_local_data(who_is_asking_file, dir_type):
@@ -6,27 +6,5 @@ def load_files_from_local_data(who_is_asking_file, dir_type):
     for key, fullpath in list_files_from_local_data(who_is_asking_file, dir_type).items():
         with open(fullpath, "rt", encoding='utf-8') as fid:
             files_dict[key] = fid.read()
-
-    return files_dict
-
-
-def list_files_from_local_data(who_is_asking_file, dir_type):
-    assert dir_type in ["repo", "user", "output"], "Argument dir_type should be either 'repo' or 'user' or 'output'."
-
-    parent_dir = os.path.dirname(os.path.realpath(who_is_asking_file))
-
-    d = {
-        "repo": "repo-files",
-        "user": "user-files",
-        "output": "snippets"
-    }[dir_type]
-
-    data_dir = os.path.join(parent_dir, d)
-    files_dict = dict()
-    for root, _, filenames in os.walk(data_dir):
-        for filename in filenames:
-            fullpath = os.path.join(root, filename)
-            key = "/".join(fullpath.replace(data_dir, "").split(os.sep))
-            files_dict[key] = fullpath
 
     return files_dict
